+++ src/core/rx_link.h
@@ -44,7 +44,7 @@ const struct rxdrv_ops *rx_link_get_ops(void);
  * Callbacks used by the link layer.
  */
 struct rx_link_cb {
-	void (*add_rx_given)(gpointer owner, gint amount);
+	void (*add_rx_given)(gpointer owner, ssize_t amount);
 	void (*read_error)(gpointer owner,
 			const gchar *reason, ...) PRINTF_FUNC_PTR(2, 3);
 	void (*got_eof)(gpointer owner);
